#391 refactor(agent): move runtime-free facade into core
