+++ packages/agent/src/core/index.ts
@@ -1,15 +1,29 @@
-export { createAgent } from '../server/createAgent'
+export { createAgent } from './createAgent'
+export type {
+  AgentCoreConfig,
+  AgentCoreRuntime,
+  AgentCoreRuntimeFactory,
+  AgentCoreRuntimeView,
+} from './createAgent'
+export type {
+  AgentCoreSessionService,
+  PiChatEventStreamResult,
+  PiChatEventStreamSubscription,
+  PiChatEventSubscriber,
+  PiChatReplayRangeError,
+  PiChatSessionService,
+  PiSessionCreateInit,
+  PiSessionRequestContext,
+} from './piChatSessionService'
 export type {
   Agent,
   AgentActor,
-  AgentConfig,
   AgentEvent,
   AgentMessageContent,
   AgentMessagePart,
   AgentReadiness,
   AgentReadinessStatus,
   AgentResolveInputResponse,
-  AgentRuntimeAdapter,
   AgentSendInput,
   AgentStartReceipt,
   AgentStreamOptions,
